requirements.txt
django
pytz
psycopg2-binary
djangorestframework
requests
pyyaml
django-filter
pytest
pytest-django
model_bakery
pytest-cov
drf-spectacular
